+++ src/presentation/controllers/signup/signup.ts
@@ -1,7 +1,6 @@
-import { MissingParamError, InvalidParamError } from '../erros'
-import { HttpRequest, HttpResponse, Controller, EmailValidator } from '../protocols'
-import { badRequest, serverError } from '../helpers/http-helper'
-import { AddAccount } from '../../domain/usecases/add-account'
+import { MissingParamError, InvalidParamError } from '../../erros'
+import { HttpRequest, HttpResponse, Controller, EmailValidator, AddAccount } from './signup-protocols'
+import { badRequest, serverError } from '../../helpers/http-helper'
 export class SignUpController implements Controller {
   constructor (private readonly emailValidator: EmailValidator, private readonly addAccount: AddAccount) {}
