+++ src/presentation/protocols/index.ts
@@ -1,3 +1,2 @@
 export * from './controller'
-export * from './email-validator'
 export * from './https'
